RESOLVED - # 125: Do not display perso module entries in standard menus if not active http://bugs.jaubart.com/view.php?id=125 RESOLVED - # 122: Hook::getNumberActiveModules method does not check whether the module is active http://bugs.jaubart.com/view.php?id=122 RESOLVED - # 123: Fix some translations http://bugs.jaubart.com/view.php?id=123
git-svn-id: http://subversion.assembla.com/svn/webtrees-geneajaubart/trunk@56 2443a68c-d498-3b47-9643-c9de34c2b0bc

+++ library/WT/Perso/Hook.php
@@ -23,7 +23,7 @@ class WT_Perso_Hook {
 	//Constants
 	private static $DEFAULT_PRIORITY = 99;
 	private static $_isModuleOperational = -1;
-	
+
 	/**
 	 * Constructor for Hook class
 	 *
@@ -47,10 +47,12 @@ class WT_Perso_Hook {
 	 * @param string $hsubscriber Name of the subscriber module
 	 */
 	public function subscribe($hsubscriber){
-		$statement = WT_DB::prepare(
-			"INSERT IGNORE INTO `##phooks` (ph_hook_function, ph_hook_context, ph_module_name)".
-			" VALUES (?, ?, ?)"
+		if(self::isModuleOperational()){
+			$statement = WT_DB::prepare(
+					"INSERT IGNORE INTO `##phooks` (ph_hook_function, ph_hook_context, ph_module_name)".
+					" VALUES (?, ?, ?)"
 			)->execute(array($this->hook_function, $this->hook_context, $hsubscriber));
+		}
 	}
 
 	/**
@@ -60,13 +62,15 @@ class WT_Perso_Hook {
 	 * @param int $priority Priority of execution
 	 */
 	public function setPriority($hsubscriber, $priority){
-		WT_DB::prepare(
+		if(self::isModuleOperational()){
+			WT_DB::prepare(
 			"UPDATE `##phooks`".
 			" SET ph_module_priority=?".
 			" WHERE ph_hook_function=?".
 			" AND ph_hook_context=?".
 			" AND ph_module_name=?"
-			)->execute(array($priority, $this->hook_function, $this->hook_context, $hsubscriber));
+					)->execute(array($priority, $this->hook_function, $this->hook_context, $hsubscriber));
+		}
 	}
 
 	/**
@@ -75,13 +79,15 @@ class WT_Perso_Hook {
 	 * @param string $hsubscriber Name of the subscriber module
 	 */
 	public function enable($hsubscriber){
+		if(self::isModuleOperational()){
 		WT_DB::prepare(
-			"UPDATE `##phooks`".
-			" SET ph_status='enabled'".
-			" WHERE ph_hook_function=?".
-			" AND ph_hook_context=?".
-			" AND ph_module_name=?"
-			)->execute(array($this->hook_function, $this->hook_context, $hsubscriber));
+		"UPDATE `##phooks`".
+		" SET ph_status='enabled'".
+		" WHERE ph_hook_function=?".
+		" AND ph_hook_context=?".
+		" AND ph_module_name=?"
+				)->execute(array($this->hook_function, $this->hook_context, $hsubscriber));
+		}
 	}
 
 	/**
@@ -90,13 +96,15 @@ class WT_Perso_Hook {
 	 * @param string $hsubscriber Name of the subscriber module
 	 */
 	public function disable($hsubscriber){
+		if(self::isModuleOperational()){
 		WT_DB::prepare(
-			"UPDATE `##phooks`".
-			" SET ph_status='disabled'".
-			" WHERE ph_hook_function=?".
-			" AND ph_hook_context=?".
-			" AND ph_module_name=?"
-			)->execute(array($this->hook_function, $this->hook_context, $hsubscriber));
+		"UPDATE `##phooks`".
+		" SET ph_status='disabled'".
+		" WHERE ph_hook_function=?".
+		" AND ph_hook_context=?".
+		" AND ph_module_name=?"
+				)->execute(array($this->hook_function, $this->hook_context, $hsubscriber));
+		}
 	}
 
 	/**
@@ -105,12 +113,14 @@ class WT_Perso_Hook {
 	 * @param string $hsubscriber Name of the subscriber module
 	 */
 	public function remove($hsubscriber){
+		if(self::isModuleOperational()){
 		WT_DB::prepare(
-			"DELETE FROM `##phooks`".
-			" WHERE ph_hook_function=?".
-			" AND ph_hook_context=?".
-			" AND ph_module_name=?"
-			)->execute(array($this->hook_function, $this->hook_context, $hsubscriber));
+		"DELETE FROM `##phooks`".
+		" WHERE ph_hook_function=?".
+		" AND ph_hook_context=?".
+		" AND ph_module_name=?"
+				)->execute(array($this->hook_function, $this->hook_context, $hsubscriber));
+		}
 	}
 
 
@@ -136,10 +146,10 @@ class WT_Perso_Hook {
 				$sqlquery = " OR ph_hook_context=?";
 			}
 			$module_names=WT_DB::prepare(
-				"SELECT ph_module_name AS module, ph_module_priority AS priority FROM `##phooks`".
-				" WHERE ph_hook_function = ? AND (ph_hook_context='all'".$sqlquery.") AND ph_status='enabled'".
-				" ORDER BY ph_module_priority ASC, module ASC"
-				)->execute($sqlparams)->fetchAssoc();
+					"SELECT ph_module_name AS module, ph_module_priority AS priority FROM `##phooks`".
+					" WHERE ph_hook_function = ? AND (ph_hook_context='all'".$sqlquery.") AND ph_status='enabled'".
+					" ORDER BY ph_module_priority ASC, module ASC"
+			)->execute($sqlparams)->fetchAssoc();
 			asort($module_names);
 			foreach ($module_names as $module_name => $module_priority) {
 				require_once WT_ROOT.WT_MODULES_DIR.$module_name.'/module.php';
@@ -150,31 +160,34 @@ class WT_Perso_Hook {
 		}
 		return $result;
 	}
-	
+
 	/*
 	 * Returns the number of active modules linked to a hook
-	 * 
-	 * @return int Nomber of active modules
-	 */
+	*
+	* @return int Number of active modules
+	*/
 	public function getNumberActiveModules(){
-		$sqlquery = '';
-		$sqlparams = array($this->hook_function);
-		if($this->hook_context != 'all') {
-			$sqlparams = array($this->hook_function, $this->hook_context);
-			$sqlquery = " OR ph_hook_context=?";
-		}
-		$module_names=WT_DB::prepare(
-			"SELECT ph_module_name AS modules FROM `##phooks`".
-			" WHERE ph_hook_function = ? AND (ph_hook_context='all'".$sqlquery.") AND ph_status='enabled'"
+		if(self::isModuleOperational()){
+			$sqlquery = '';
+			$sqlparams = array($this->hook_function);
+			if($this->hook_context != 'all') {
+				$sqlparams = array($this->hook_function, $this->hook_context);
+				$sqlquery = " OR ph_hook_context=?";
+			}
+			$module_names=WT_DB::prepare(
+					"SELECT ph_module_name AS modules FROM `##phooks`".
+					" WHERE ph_hook_function = ? AND (ph_hook_context='all'".$sqlquery.") AND ph_status='enabled'"
 			)->execute($sqlparams)->fetchOneColumn();
-		return count($module_names);
+			return count($module_names);
+		}
+		return 0;
 	}
-	
+
 	/*
 	 * Return whether any active module is linked to a hook
-	 * 
-	 * @return bool True is active modules exist, false otherwise
-	 */
+	*
+	* @return bool True is active modules exist, false otherwise
+	*/
 	public function hasAnyActiveModule(){
 		return ($this->getNumberActiveModules()>0);
 	}
@@ -182,9 +195,9 @@ class WT_Perso_Hook {
 	/**
 	 * Static functions for Hooks.
 	 */
-	
+
 	/**
-	 * Return whether the Hook module is active and the table has been created. 
+	 * Return whether the Hook module is active and the table has been created.
 	 *
 	 * @return bool True if module active and table created, false otherwise
 	 */
@@ -249,18 +262,21 @@ class WT_Perso_Hook {
 		}
 		return $hooks;
 	}
-	
+
 	/**
 	 * Get the list of hooks intalled in webtrees, with their id, status and priority.
 	 *
 	 * @return DBOStatement List of installed hooks
 	 */
 	static public function getRawInstalledHooks(){
+		if(self::isModuleOperational()){
 		return WT_DB::prepare(
-			"SELECT ph_id AS id, ph_module_name AS module, ph_hook_function AS hook, ph_hook_context as context, ph_module_priority AS priority,  ph_status AS status".
-			" FROM `##phooks`".
-			" ORDER BY hook ASC, status ASC, priority ASC, module ASC"
+				"SELECT ph_id AS id, ph_module_name AS module, ph_hook_function AS hook, ph_hook_context as context, ph_module_priority AS priority,  ph_status AS status".
+				" FROM `##phooks`".
+				" ORDER BY hook ASC, status ASC, priority ASC, module ASC"
 		)->execute()->fetchAll();
+		}
+		return array();
 	}
 
 	/**
